microservices/document-extractor/code/requirements.txt
Flask==2.3.3
pypdf==6.6.0
python-docx==1.2.0
